refactor: organización de la estructura del proyecto
