+++ lib/presentation/providers/chat_provider.dart
@@ -1,12 +1,11 @@
-import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:yes_no_app/domain/entities/message.dart';
 import 'package:yes_no_app/infrastructure/services/bible_service.dart';
-import 'package:yes_no_app/infrastructure/models/bible_verse_model.dart';
+import 'package:yes_no_app/data/models/bible_verse_model.dart';
 
 /// Estados possíveis durante o fluxo de busca de versículos
 enum SearchState {
-  initial,      // Estado inicial, pronto para receber uma busca
+  initial, // Estado inicial, pronto para receber uma busca
   waitingChoice, // Esperando que o usuário escolha um versículo da lista
   waitingNewSearch, // Esperando confirmação para iniciar uma nova busca
 }
@@ -21,7 +20,7 @@ class ChatProvider extends ChangeNotifier {
   final bool _isJsonLoaded = true;
   String _userName = '';
   BuildContext? _context;
-  
+
   ChatProvider({bool initializeChat = true}) {
     if (initializeChat) {
       // Iniciamos la carga de mensajes iniciales después de que el widget esté construido
@@ -32,7 +31,7 @@ class ChatProvider extends ChangeNotifier {
       });
     }
   }
-  
+
   /// Define o nome do usuário
   void setContext(BuildContext context) {
     _context = context;
@@ -57,7 +56,7 @@ class ChatProvider extends ChangeNotifier {
       'Escribe lo que tienes en mente, por ejemplo: "amor", "perdón" o "Salmo 23:1".',
       'Si quieres salir del chat en cualquier momento, puedes escribir "SALIR".',
     ];
-    
+
     for (final message in initialMessages) {
       await _showTypingThenMessage(message);
       await Future.delayed(const Duration(milliseconds: 300));
@@ -67,16 +66,16 @@ class ChatProvider extends ChangeNotifier {
   /// Adiciona a mensagem diretamente sem efeito de digitação
   Future<void> _showTypingThenMessage(dynamic messageContent) async {
     if (messageContent is String) {
-      _addMessage(Message(
-        text: messageContent,
-        fromWho: FromWho.systemChatMessage,
-      ));
+      _addMessage(
+        Message(text: messageContent, fromWho: FromWho.systemChatMessage),
+      );
     } else if (messageContent is Message) {
       _addMessage(messageContent);
     }
-    
+
     await moveScrollToBottom();
   }
+
   // Getters
   bool get isWaitingChoice => _currentState == SearchState.waitingChoice;
   bool get isWaitingNewSearch => _currentState == SearchState.waitingNewSearch;
@@ -85,38 +84,38 @@ class ChatProvider extends ChangeNotifier {
   /// Processa a mensagem enviada pelo usuário
   Future<void> sendMessage(String text) async {
     if (text.isEmpty) return;
-    
+
     // Verificar SALIR antes de cualquier otra operación
     if (text.toUpperCase().trim() == 'SALIR') {
       _addUserChatMessage(text);
       // Mostrar mensajes de despedida
       await _addSystemMessages([
         '¡Hasta luego!',
-        'Gracias por usar el asistente bíblico.'
+        'Gracias por usar el asistente bíblico.',
       ]);
       await Future.delayed(const Duration(seconds: 2));
-      
+
       // Limpiar completamente el chat
       messageList.clear();
       notifyListeners();
-      
+
       // Esperar un momento para que se vea la limpieza
       await Future.delayed(const Duration(milliseconds: 500));
-      
+
       // Reiniciar el estado
       _resetSearch();
       _currentState = SearchState.initial;
-      
+
       // Redirigir a la pantalla inicial y limpiar el historial de navegación
       if (_context != null) {
         Navigator.of(_context!).pushNamedAndRemoveUntil(
           '/',
-          (route) => false,  // Esto removerá todas las rutas anteriores
+          (route) => false, // Esto removerá todas las rutas anteriores
         );
       }
       return;
     }
-    
+
     if (!_isJsonLoaded) {
       await _addSystemChatMessage("Cargando versos, por favor espere...");
       return;
@@ -133,60 +132,67 @@ class ChatProvider extends ChangeNotifier {
         await _searchVerses(text);
       }
     } catch (e) {
-      await _addSystemChatMessage('Ocurrió un error. Por favor, intenta de nuevo.');
+      await _addSystemChatMessage(
+        'Ocurrió un error. Por favor, intenta de nuevo.',
+      );
     }
 
     await moveScrollToBottom();
   }
 
   /// Processa a escolha do usuário
-Future<void> _processChoice(String text) async {
+  Future<void> _processChoice(String text) async {
     final upperText = text.toUpperCase().trim();
-    final invalidOptionsMessage = 'Escribe "BUSCAR" para una nueva búsqueda.\nEscribe "VOLVER" para regresar a los versículos.\nEscribe "SALIR" si quieres terminar el chat.';
-    
+    final invalidOptionsMessage =
+        'Escribe "BUSCAR" para una nueva búsqueda.\nEscribe "VOLVER" para regresar a los versículos.\nEscribe "SALIR" si quieres terminar el chat.';
+
     switch (upperText) {
-        case 'BUSCAR':
-            _resetSearch();
-            await _addSystemChatMessage('Escribe lo que tengas en mente, por ejemplo: "amor", "perdón" o "Salmos 23:1".');
-            _currentState = SearchState.initial;
-            return;
-            
-        case 'VOLVER':
-            await _showVersesList();
-            _currentState = SearchState.waitingChoice;
-            return;
-            
-        case 'SALIR':
-            // La lógica de salida ahora está en sendMessage
-            return;
+      case 'BUSCAR':
+        _resetSearch();
+        await _addSystemChatMessage(
+          'Escribe lo que tengas en mente, por ejemplo: "amor", "perdón" o "Salmos 23:1".',
+        );
+        _currentState = SearchState.initial;
+        return;
+
+      case 'VOLVER':
+        await _showVersesList();
+        _currentState = SearchState.waitingChoice;
+        return;
+
+      case 'SALIR':
+        // La lógica de salida ahora está en sendMessage
+        return;
     }
 
     final choice = int.tryParse(text);
     if (_isValidChoice(choice)) {
-        await _showSelectedVerse(choice!);
-        // Agregamos una pausa más corta para mejorar la fluidez
-        await Future.delayed(const Duration(milliseconds: 300));
-        _currentState = SearchState.waitingChoice;
-        await _showVersesList();
+      await _showSelectedVerse(choice!);
+      // Agregamos una pausa más corta para mejorar la fluidez
+      await Future.delayed(const Duration(milliseconds: 300));
+      _currentState = SearchState.waitingChoice;
+      await _showVersesList();
     } else {
-        await _addSystemMessages([
-            '¡Ups! Esa opción no es válida.',
-            'Por favor, elige una opción:',
-            invalidOptionsMessage
-        ]);
-        _currentState = SearchState.waitingChoice;
+      await _addSystemMessages([
+        '¡Ups! Esa opción no es válida.',
+        'Por favor, elige una opción:',
+        invalidOptionsMessage,
+      ]);
+      _currentState = SearchState.waitingChoice;
     }
-}
+  }
 
   /// Processa a resposta para uma nova busca
   Future<void> _handleNewSearchResponse(String text) async {
     final response = text.toLowerCase().trim();
-    
+
     if (response == 'salir') {
       await _exitChat();
     } else if (response == 'si' || response == 'sí') {
       _resetSearch();
-      await _addSystemChatMessage('Escribe lo que tengas en mente, por ejemplo: "amor", "perdón" o "Salmos 23:1".');
+      await _addSystemChatMessage(
+        'Escribe lo que tengas en mente, por ejemplo: "amor", "perdón" o "Salmos 23:1".',
+      );
       _currentState = SearchState.initial;
     } else if (response == 'no') {
       await _showVersesList();
@@ -195,7 +201,7 @@ Future<void> _processChoice(String text) async {
       await _addSystemChatMessage('Por favor, responda "SÍ", "NO" o "SALIR".');
       _currentState = SearchState.waitingNewSearch;
     }
-    
+
     notifyListeners();
   }
 
@@ -203,28 +209,25 @@ Future<void> _processChoice(String text) async {
   Future<void> _exitChat() async {
     await _addSystemMessages([
       '¡Hasta luego!',
-      'Gracias por usar el asistente bíblico.'
+      'Gracias por usar el asistente bíblico.',
     ]);
     await Future.delayed(const Duration(seconds: 2));
-    
+
     // Limpiamos completamente el chat
     messageList.clear();
     notifyListeners();
-    
+
     // Esperamos un momento para que se vea la limpieza
     await Future.delayed(const Duration(milliseconds: 500));
-    
+
     // Reiniciamos el estado
     _resetSearch();
     _currentState = SearchState.initial;
     _userName = ''; // Limpiamos también el nombre de usuario
-    
+
     // Redirigir a la pantalla inicial y limpiar el historial de navegación
     if (_context != null) {
-      Navigator.of(_context!).pushNamedAndRemoveUntil(
-        '/',
-        (route) => false,
-      );
+      Navigator.of(_context!).pushNamedAndRemoveUntil('/', (route) => false);
     }
   }
 
@@ -242,10 +245,7 @@ Future<void> _processChoice(String text) async {
   }
 
   void _addUserChatMessage(String text) {
-    _addMessage(Message(
-      text: text,
-      fromWho: FromWho.userChatMessage,
-    ));
+    _addMessage(Message(text: text, fromWho: FromWho.userChatMessage));
   }
 
   void _addMessage(Message message) {
@@ -270,9 +270,7 @@ Future<void> _processChoice(String text) async {
 
   /// Mostra um versículo selecionado pelo usuário
   Future<void> _showSelectedVerse(int choice) async {
-    await _showTypingThenMessage(
-      _searchResults[choice - 1].toMessageEntity()
-    );
+    await _showTypingThenMessage(_searchResults[choice - 1].toMessageEntity());
   }
 
   void _resetSearch() {
@@ -288,10 +286,14 @@ Future<void> _processChoice(String text) async {
   }
 
   String _createVersesList() {
-    final buffer = StringBuffer("Encontré ${_searchResults.length} versículos:\n\n");
+    final buffer = StringBuffer(
+      "Encontré ${_searchResults.length} versículos:\n\n",
+    );
     for (var i = 0; i < _searchResults.length; i++) {
       final verse = _searchResults[i];
-      buffer.write("${i + 1}. ${verse.livro} ${verse.capitulo}:${verse.versiculo}\n");
+      buffer.write(
+        "${i + 1}. ${verse.livro} ${verse.capitulo}:${verse.versiculo}\n",
+      );
     }
     return buffer.toString();
   }
@@ -300,13 +302,13 @@ Future<void> _processChoice(String text) async {
   Future<void> moveScrollToBottom() async {
     try {
       await Future.delayed(const Duration(milliseconds: 100));
-      
+
       WidgetsBinding.instance.addPostFrameCallback((_) {
         if (!chatScrollController.hasClients) return;
-        
+
         final maxScroll = chatScrollController.position.maxScrollExtent;
         final currentScroll = chatScrollController.position.pixels;
-        
+
         if (currentScroll < maxScroll) {
           chatScrollController.animateTo(
             maxScroll,
