+++ test/infrastructure/services/bible_service_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_test/flutter_test.dart';
 import 'package:yes_no_app/infrastructure/services/bible_service.dart';
-import 'package:yes_no_app/infrastructure/models/bible_verse_model.dart';
-import 'package:yes_no_app/infrastructure/repositories/bible_repository.dart';
+import 'package:yes_no_app/data/models/bible_verse_model.dart';
+import 'package:yes_no_app/domain/repositories/bible_repository.dart';
 
 // Mock do BibleRepository
 class MockBibleRepository implements IBibleRepository {
